feat: Result타입 NetworkError를 활용하여 데이터와 모델 연결 시 에러핸들링 구현
