+++ WeatherForecast/WeatherForecast/ Model/CurrentWeatherInformation.swift
