+++ WeatherForecast/WeatherForecast/ Model/FiveDaysForecastingInformation.swift
